move jwt auth to auth module
