+++ core/amber/src/main/scala/edu/uci/ics/texera/web/ServletAwareConfigurator.scala
@@ -1,7 +1,7 @@
 package edu.uci.ics.texera.web
 
 import com.typesafe.scalalogging.LazyLogging
-import edu.uci.ics.texera.web.auth.JwtAuth.jwtConsumer
+import edu.uci.ics.texera.auth.JwtAuth.jwtConsumer
 import edu.uci.ics.texera.dao.jooq.generated.tables.pojos.User
 import org.apache.http.client.utils.URLEncodedUtils
